finished Diffie-Hellman without input sanitization and handling

+++ DiffieHellman.java
@@ -1,5 +1,5 @@
 public class DiffieHellman {
-	public static int calculateKey(int g, int a, int p) {
-		return (int) (Math.pow(g, a) % p);
+	public static long calculateKey(long g, long a, long p) {
+		return (long) (Math.pow(g, a) % p);
 	}
 }
